[Plugin] Rename OAuth.js -> oauth.js
